Internal: Conversion Center - Introduce trait and configuration (#25820)
