+++ assets/dev/scss/frontend/_modules.scss
@@ -2,3 +2,4 @@
 @import '../../../../modules/shapes/assets/scss/frontend';
 @import '../../../../modules/nested-tabs/assets/scss/frontend';
 @import '../../../../modules/nested-accordion/assets/scss/frontend';
+@import '../../../../modules/conversion-center/assets/scss/frontend';
