+++ assets/dev/scss/frontend/_widgets.scss
@@ -24,4 +24,3 @@
 @import "widgets/toggle";
 @import "widgets/video";
 @import "widgets/rating";
-@import "widgets/link-in-bio";
