+++ modules/conversion-center/assets/scss/widgets/link-in-bio-base.scss
@@ -1,7 +1,4 @@
 .e-link-in-bio {
-	$link-in-bio-cta-bg-color: #467FF7;
-	$link-in-bio-text-color-alt: #324a6d;
-	$link-in-bio-text-color-default: #1c2448;
 
 	// Layout config
 	--e-link-in-bio-border-color: transparent;
@@ -37,12 +34,12 @@
 	// Icons config
 	--e-link-in-bio-icon-color: inherit;
 	--e-link-in-bio-icon-gap: 20px 29px;
-	--e-link-in-bio-icon-size: 25px;
+	--e-link-in-bio-icon-size: #{$link-in-bio-icon-size-small};
 
 	// CTAs config
 	--e-link-in-bio-ctas-background-color: #{$link-in-bio-cta-bg-color};
 	--e-link-in-bio-ctas-border-color: transparent;
-	--e-link-in-bio-ctas-border-radius: 20px;
+	--e-link-in-bio-ctas-border-radius: #{$link-in-bio-border-radius-rounded};
 	--e-link-in-bio-ctas-border-style: none;
 	--e-link-in-bio-ctas-border-width: 2px;
 	--e-link-in-bio-ctas-gap: 22px;
@@ -237,11 +234,11 @@
 		height: var(--e-link-in-bio-icon-size);
 
 		&.has-size-medium {
-			--e-link-in-bio-icon-size: 30px;
+			--e-link-in-bio-icon-size: #{$link-in-bio-icon-size-medium};
 		}
 
 		&.has-size-large {
-			--e-link-in-bio-icon-size: 35px;
+			--e-link-in-bio-icon-size: #{$link-in-bio-icon-size-large};
 		}
 
 		.e-link-in-bio__icon-link {
@@ -270,6 +267,8 @@
 
 		&.has-type-link {
 			--e-link-in-bio-ctas-gap: 10px;
+
+			justify-items: center;
 		}
 
 		.e-link-in-bio__cta {
@@ -278,7 +277,6 @@
 			font-size: 16px;
 			font-weight: 500;
 			line-height: 20px;
-			justify-content: center;
 
 			&:hover,
 			&:focus,
@@ -287,13 +285,27 @@
 			}
 		}
 
+		.e-link-in-bio__cta-image {
+			flex: 0 0 min(50%, 140px);
+		}
+
+		.e-link-in-bio__cta-image-el {
+			aspect-ratio: 140 / 100;
+			height: 100%;
+			object-fit: cover;
+			width: 100%;
+		}
+
+		.e-link-in-bio__cta-text {
+			align-items: center;
+			display: flex;
+			flex: 1 1 auto;
+			justify-content: center;
+		}
+
 		.e-link-in-bio__cta.is-type-button {
 			border-radius: var(--e-link-in-bio-ctas-border-radius);
-			background-color: var(--e-link-in-bio-ctas-background-color);
-			padding-block-end: var(--e-link-in-bio-ctas-padding-block-end);
-			padding-block-start: var(--e-link-in-bio-ctas-padding-block-start);
-			padding-inline-end: var(--e-link-in-bio-ctas-padding-inline-end);
-			padding-inline-start: var(--e-link-in-bio-ctas-padding-inline-start);
+			overflow: hidden;
 
 			&.has-border {
 				--e-link-in-bio-ctas-border-style: solid;
@@ -302,12 +314,24 @@
 				border-width: var(--e-link-in-bio-ctas-border-width);
 			}
 
+			&.has-corners-rounded {
+				--e-link-in-bio-ctas-border-radius: #{$link-in-bio-border-radius-rounded};
+			}
+
 			&.has-corners-round {
-				--e-link-in-bio-ctas-border-radius: 50px;
+				--e-link-in-bio-ctas-border-radius: #{$link-in-bio-border-radius-round};
 			}
 
 			&.has-corners-sharp {
-				--e-link-in-bio-ctas-border-radius: 0;
+				--e-link-in-bio-ctas-border-radius: #{$link-in-bio-border-radius-sharp};
+			}
+
+			.e-link-in-bio__cta-text {
+				background-color: var(--e-link-in-bio-ctas-background-color);
+				padding-block-end: var(--e-link-in-bio-ctas-padding-block-end);
+				padding-block-start: var(--e-link-in-bio-ctas-padding-block-start);
+				padding-inline-end: var(--e-link-in-bio-ctas-padding-inline-end);
+				padding-inline-start: var(--e-link-in-bio-ctas-padding-inline-start);
 			}
 		}
 
@@ -316,6 +340,7 @@
 
 			font-weight: 700;
 			margin-block: 17px;
+			justify-content: center;
 		}
 	}
