+++ modules/conversion-center/assets/scss/widgets/link-in-bio-var-3.scss
@@ -1,0 +1,28 @@
+.e-link-in-bio-var-3 {
+	--e-link-in-bio-gutter-block-end: 20px;
+	--e-link-in-bio-ctas-border-radius: #{$link-in-bio-border-radius-sharp};
+	--e-link-in-bio-ctas-gap: 8px;
+
+	.e-link-in-bio__ctas {
+
+		.e-link-in-bio__cta {
+
+			// Move border to image
+			&.has-border {
+				border: none;
+
+				.e-link-in-bio__cta-image {
+					--e-link-in-bio-ctas-border-style: solid;
+					border-color: var(--e-link-in-bio-ctas-border-color);
+					border-style: var(--e-link-in-bio-ctas-border-style);
+					border-width: var(--e-link-in-bio-ctas-border-width);
+				}
+			}
+
+			&.has-corners-rounded,
+			&.has-corners-round {
+				--e-link-in-bio-ctas-border-radius: #{$link-in-bio-border-radius-sharp};
+			}
+		}
+	}
+}
