+++ modules/conversion-center/assets/scss/widgets/link-in-bio-vars.scss
@@ -1,0 +1,12 @@
+// Constants
+$link-in-bio-cta-bg-color: #467FF7;
+$link-in-bio-text-color-alt: #324a6d;
+$link-in-bio-text-color-default: #1c2448;
+
+$link-in-bio-border-radius-rounded: 20px;
+$link-in-bio-border-radius-round: 50px;
+$link-in-bio-border-radius-sharp: 0;
+
+$link-in-bio-icon-size-small: 25px;
+$link-in-bio-icon-size-medium: 30px;
+$link-in-bio-icon-size-large: 35px;
